--- backend/src/main/kotlin/be/osoc/team1/backend/controllers/PositionController.kt
@@ -1,10 +1,0 @@
-package be.osoc.team1.backend.controllers
-
-import be.osoc.team1.backend.entities.Position
-import be.osoc.team1.backend.services.PositionService
-import org.springframework.web.bind.annotation.RequestMapping
-import org.springframework.web.bind.annotation.RestController
-
-@RestController
-@RequestMapping("/positions")
-class PositionController(service: PositionService) : BaseController<Position>(service)
